requirements.txt
fastapi
uvicorn
trafilatura
pydantic
google-genai
jinja2
python-dotenv
python-multipart